Carne refeito e quase finalizado

+++ projeto/emprestimo/models.py
@@ -11,9 +11,9 @@ class Emprestimo(models.Model):
     funcionario      = models.ForeignKey(User, on_delete=models.CASCADE)
     cliente          = models.ForeignKey(Cliente, on_delete=models.CASCADE, null=True, blank=True)
     num_doc          = models.PositiveIntegerField('Nº Documento', null=True, blank=True)    
-    valor_emprestado = models.DecimalField("Valor Emprestado", max_digits=10, decimal_places=2, null=True, blank=True)
+    valor_emprestado = models.FloatField("Valor Emprestado", null=True, blank=True)
     qtd_parcelas     = models.PositiveIntegerField('Qtd Parcelas', null=True, blank=True)
-    valor_prestacao  = models.DecimalField("Valor da prestação", max_digits=10, decimal_places=2, null=True, blank=True)
+    valor_prestacao  = models.FloatField("Valor da prestação", null=True, blank=True)
     data_emprestimo  = models.DateField("Data do Emprestimo", auto_now_add=False, auto_now=False, null=True,blank=True,)    
 
     class Meta:
